refactor: extract FundamentalsPage data logic into useFundamentalsData hook (F5)
Move FundamentalsPage's fetch/state/effects into a page-local hook
(components/fundamentals/useFundamentalsData.js) with its hook-only helpers
(tab configs, market normalizers, isUsAsset), and the shared MARKET_OPTIONS
config into fundamentals/fundamentalsUtils.js. The page becomes a render
shell (858 -> 631 lines) keeping only its JSX and the row-config data it
renders.

Behavior-preserving: hook body moved verbatim (same hook order/effect deps),
returning exactly the state and handlers the JSX consumes; no early returns
in the body. Verified by FundamentalsPage's existing tests, full suite (83),
lint, and build.
