+++ frontend/src/components/fundamentals/fundamentalsUtils.js
@@ -1,0 +1,7 @@
+// Shared config for the fundamentals page and its data hook.
+
+export const MARKET_OPTIONS = [
+    { value: 'us', label: 'US' },
+    { value: 'hk', label: 'HK' },
+    { value: 'cn', label: 'CN' },
+];
